fix(community): update community router and API files
